backend/requirements.txt
django==4.*
python-dotenv
psycopg2-binary
djangorestframework-simplejwt
django-cors-headers
